panier: resolution pb clé hashmap en mettant une variable static qui s'incremente a chaque instanciation d'un objet LigneCommande.

+++ bahia_beach/bahia_beach-ejb/src/java/beanMetier/beanPanierLocal.java
@@ -17,7 +17,7 @@ public interface beanPanierLocal {
 
     public void add(Long id);
 
-    public void delete(Long id);
+    public void delete(int id);
 
     public void clearPanier();
